backend/requirements.txt
# FastAPI et serveur
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21

# Database
sqlalchemy==2.0.45
pyodbc==5.3.0

# Configuration
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# Utilities
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.2
Pillow>=10.3.0
